reparando comentarios

+++ app/Http/Controllers/InstanceController.php
@@ -35,7 +35,7 @@ class InstanceController extends Controller
      */
     public function create($record)
     {
-       // $this->authorize('admin');
+        $this->authorize('admin');
         $record = Record::find($record, ['id','titulo']);
         return view('instance.create')->with(['record' => $record]);
     }
@@ -48,7 +48,7 @@ class InstanceController extends Controller
      */
     public function store(InstanceRequest $request)
     {
-       // $this->authorize('admin');
+        $this->authorize('admin');
         $validated = $request->validated();
 
         Instance::create($validated);
@@ -64,7 +64,7 @@ class InstanceController extends Controller
      */
     public function show($instance)
     {
-       // $this->authorize('admin');
+        $this->authorize('admin');
         $instance = Instance::with('record:id,titulo')->find($instance);
         return view('instance.show')->with([
             'instance' => $instance,
@@ -79,7 +79,7 @@ class InstanceController extends Controller
      */
     public function edit($instance)
     {
-       // $this->authorize('admin');
+        $this->authorize('admin');
         $instance = Instance::with('record:id,titulo')->find($instance);
         return view('instance.edit')->with([
             'instance' => $instance
@@ -95,7 +95,7 @@ class InstanceController extends Controller
      */
     public function update(InstanceRequest $request, Instance $instance)
     {
-       // $this->authorize('admin');
+        $this->authorize('admin');
         $validated = $request->validated();
         $instance->update($validated);
 
@@ -110,7 +110,7 @@ class InstanceController extends Controller
      */
     public function destroy(Instance $instance)
     {
-       // $this->authorize('admin');
+        $this->authorize('admin');
         $instance->delete();
         return redirect('/instance');
     }
